build(devs-infra): allow extending TsJestTransformer
Allow end users to extend `TsJestTransformer` with class inheritance
